ONNXToTorch/ConvertONNXToTorch.cpp: Update type converter
